+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next release
 
 - feat(v0.8.0-rc0): starknet_subscribeNewHeads
+- fix(rocksdb): update max open files opt
+- code: refactor to use otel tracing instead of prometheus (removed mc-metrics, added mc-analytics)
 - fix(version constants): 0.13.2 was mapped to wrong constants
 - fix(compilation): devnet contract artifacts are not compiled by `cargo build` anymore
 - feat: add fgw get_block_traces